DRYed up and organized the code a bit
