+++ app/lib/backends/opennebula/networkinterface.rb
@@ -7,7 +7,9 @@ module Backends
       include Backends::Helpers::AttributesTransferable
       include Backends::Helpers::MixinsAttachable
       include Backends::Helpers::ErbRenderer
-      include Backends::Opennebula::Helpers::Waiter
+
+      # :nodoc:
+      HELPER_NS = Backends::Opennebula::Helpers
 
       class << self
         # @see `served_class` on `Entitylike`
@@ -59,10 +61,10 @@ module Backends
         nics = Backends::Opennebula::Helpers::Counter.xml_elements(vm, 'TEMPLATE/NIC')
 
         client(Errors::Backend::EntityCreateError) { vm.nic_attach nic_from(instance, vm) }
-        wait_until(vm, 'RUNNING') do |nvm|
-          unless Backends::Opennebula::Helpers::Counter.xml_elements(nvm, 'TEMPLATE/NIC') > nics
+        HELPER_NS::Waiter.wait_until(vm, 'RUNNING') do |nvm|
+          unless HELPER_NS::Counter.xml_elements(nvm, 'TEMPLATE/NIC') > nics
             logger.error "Attaching VNET to VM[#{vm['ID']}] failed: #{vm['USER_TEMPLATE/ERROR']}"
-            raise Errors::Backend::EntityCreateError, 'Could not attach network to compute'
+            raise Errors::Backend::RemoteError, 'Could not attach network to compute'
           end
         end
