+++ app/lib/errors/backend/remote_error.rb
@@ -1,0 +1,5 @@
+module Errors
+  module Backend
+    class RemoteError < StandardError; end
+  end
+end
